+++ src/app/(site-layout)/(protected)/(me)/me/films/page.tsx
@@ -1,4 +1,4 @@
-import { UserFilmsTabs } from '@/components/user-films-tabs'
+import { UserFilmsTabs } from '@/app/(site-layout)/(protected)/(me)/me/films/components/user-films-tabs'
 import { db } from '@/lib/db'
 import { getCurrentUser } from '@/lib/session'
 import { notFound } from 'next/navigation'
