+++ src/app/(site-layout)/(public)/members/[username]/list/[title]/page.tsx
@@ -1,4 +1,4 @@
-import { UserFilmCard } from '@/components/user-film-card'
+import { UserFilmCard } from '@/components/film/user-film-card'
 import { searchByID } from '@/lib/tmdb/src/tmdb'
 import listsModel from '@/models/lists'
